+++ domain/src/main/java/ci/nsu/mobile/domain/interfaces/CalculationsProvider.kt
@@ -4,7 +4,7 @@ import ci.nsu.mobile.domain.model.DepositCalculation
 import kotlinx.coroutines.flow.Flow
 
 interface CalculationsProvider {
-    fun getCalculationsForUser(userId: Long): Flow<List<DepositCalculation>>
+    fun getCalculationsForUser(userId: Int): Flow<List<DepositCalculation>>
     suspend fun saveCalculation(calculation: DepositCalculation)
-    suspend fun deleteCalculation(calculationId: Long)
+    suspend fun deleteCalculation(calculationId: Int)
 }
